requirements.txt
altair==5.4.1 ; python_version >= "3.12" and python_version < "4.0"
attrs==24.2.0 ; python_version >= "3.12" and python_version < "4.0"
babel==2.16.0 ; python_version >= "3.12" and python_version < "4.0"
blinker==1.8.2 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
decorator==5.1.1 ; python_version >= "3.12" and python_version < "4.0"
duckdb==1.1.1 ; python_version >= "3.12" and python_version < "4.0"
gitdb==4.0.11 ; python_version >= "3.12" and python_version < "4.0"
gitpython==3.1.43 ; python_version >= "3.12" and python_version < "4.0"
google-auth-oauthlib==1.2.1 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.35.0 ; python_version >= "3.12" and python_version < "4.0"
gspread-dataframe==4.0.0 ; python_version >= "3.12" and python_version < "4.0"
gspread-formatting==1.2.0 ; python_version >= "3.12" and python_version < "4.0"
gspread-pandas==3.3.0 ; python_version >= "3.12" and python_version < "4.0"
gspread==5.12.4 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
jsonschema-specifications==2024.10.1 ; python_version >= "3.12" and python_version < "4.0"
jsonschema==4.23.0 ; python_version >= "3.12" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==3.0.1 ; python_version >= "3.12" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.12" and python_version < "4.0"
narwhals==1.9.1 ; python_version >= "3.12" and python_version < "4.0"
numpy==2.1.2 ; python_version >= "3.12" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.12" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
pillow==10.4.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.28.2 ; python_version >= "3.12" and python_version < "4.0"
pyarrow==17.0.0 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pydeck==0.9.1 ; python_version >= "3.12" and python_version < "4.0"
pygments==2.18.0 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.12" and python_version < "4.0"
referencing==0.35.1 ; python_version >= "3.12" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rich==13.9.2 ; python_version >= "3.12" and python_version < "4.0"
rpds-py==0.20.0 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
smmap==5.0.1 ; python_version >= "3.12" and python_version < "4.0"
sql-metadata==2.13.0 ; python_version >= "3.12" and python_version < "4.0"
sqlparse==0.5.1 ; python_version >= "3.12" and python_version < "4.0"
st-gsheets-connection==0.1.0 ; python_version >= "3.12" and python_version < "4.0"
streamlit==1.39.0 ; python_version >= "3.12" and python_version < "4.0"
tenacity==9.0.0 ; python_version >= "3.12" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.12" and python_version < "4.0"
tornado==6.4.1 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
tzdata==2024.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
validators==0.34.0 ; python_version >= "3.12" and python_version < "4.0"
watchdog==5.0.3 ; python_version >= "3.12" and python_version < "4.0" and platform_system != "Darwin"
